[Technical] Rework definitions of new pseudo-flavors
Rick Macklem and Dave Noveck objected to the definition of new RPC
auth flavors. Instead, they felt the RPC layer should not be involved
in authorizing the use of transport layer security with AUTH_NONE and
AUTH_SYS. Rather, the Upper-Layer Protocol would make those decisions
and return an error if appropriate security was not in place.

This patch replaces the definitions of new RPC auth flavors with
definitions of true pseudo-flavors, and updates associated text in
Section 3 to align with this new direction.