[Technical] Update example showing how NFSv3 uses the new p-flavor
Replace the use of pseudo-flavors with the use of a traditional RPC
auth flavor and channel binding, as requested by Rick Macklem and
Dave Noveck.

As part of the update, include discussion of NFSv2 for completeness.

Provide an appropriate NFS status code that legacy NFS servers can
use to report a security policy violation.